[Chore]: Add form component, change props type
